# specs/001-comidapp-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ComidApp API
  version: 1.0.0
  description: >
    API REST para ComidApp — plataforma web de pedidos y entrega de comida rápida.
    Todos los endpoints protegidos requieren token JWT en el header Authorization.
  contact:
    name: ComidApp Team
servers:
  - url: http://localhost:3001/api/v1
    description: Desarrollo local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RolUsuario:
      type: string
      enum: [CLIENTE, ADMIN, REPARTIDOR]

    # ── Shared ───────────────────────────────────────────────────────────────
...
    UsuarioPublico:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
          format: email
        telefono:
          type: string
          nullable: true
        direccionEntrega:
          type: string
          nullable: true
        rol:
          $ref: '#/components/schemas/RolUsuario'

    # ── Locales ──────────────────────────────────────────────────────────────
...
paths:
...
  /usuarios/perfil:
    get:
      tags: [Usuarios]
      summary: Obtener perfil del usuario autenticado (FR-021)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioPublico'

    patch:
      tags: [Usuarios]
      summary: Actualizar datos del perfil (FR-021)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                telefono:
                  type: string
                direccionEntrega:
                  type: string
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioPublico'
...